requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
httpx==0.27.2
pydantic==2.9.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
